+++ js/create-dom-element.js
@@ -1,0 +1,11 @@
+export const creatDOMElement = (markup, className) => {
+  let newElement;
+
+  newElement = document.createElement(`section`);
+
+  newElement.className = className;
+  newElement.innerHTML = markup;
+
+  return newElement;
+};
+
