+++ js/mistakes.js
@@ -1,8 +1,9 @@
 export const mistakes = (mistakesNumber) => {
-  return `<div class="main-mistakes">
+  return `
+    <div class="main-mistakes">
     ${new Array(mistakesNumber)
       .fill(`<img class="main-mistake" src="img/wrong-answer.png" width="35" height="49">`)
       .join(``)}
-  </div>
-`;
+    </div>
+  `;
 };
