+++ js/show-page.js
@@ -1,0 +1,6 @@
+const app = document.querySelector(`.app`);
+
+export const showPage = (page) => {
+  const newPage = page();
+  app.replaceChild(newPage.element, app.querySelector(`.main`));
+};
