+++ js/show-random-page.js
@@ -4,6 +4,7 @@ import {currentGame} from '../js/game-store.js';
 import {levelGenre} from '../js/level-genre.js';
 import {levelArtist} from '../js/level-artist.js';
 import {finalResult} from '../js/result.js';
+import {showPage} from '../js/show-page.js';
 
 const ARTIST_PAGE_INDEX = 1;
 const GENRE_PAGE_INDEX = 2;
@@ -11,18 +12,18 @@ const LAST_INDEX = 11;
 
 export const showRandomPage = () => {
   if (currentGame.state.lives === MINIMUM_PLAYERS_LIVES) {
-    return finalResult().element;
+    showPage(finalResult);
   } else if (currentGame.state.round < LAST_INDEX) {
     currentGame.nextRound();
 
     let nextPageIndex = getRandomInt(ARTIST_PAGE_INDEX, GENRE_PAGE_INDEX);
 
     if (nextPageIndex === ARTIST_PAGE_INDEX) {
-      return levelArtist().element;
+      showPage(levelArtist);
     } else {
-      return levelGenre().element;
+      showPage(levelGenre);
     }
   } else {
-    return finalResult().element;
+    showPage(finalResult);
   }
 };
